# exeuctor: fix data race for TestPointGetLockExistKey