--- a/executor/point_get_test.go
+++ b/executor/point_get_test.go
@@ -18,6 +18,7 @@ import (
 	"context"
 	"fmt"
 	"strings"
+	"sync"
 	"testing"
 	"time"
 
@@ -683,8 +684,6 @@ func TestPointGetWriteLock(t *testing.T) {
 }
 
 func TestPointGetLockExistKey(t *testing.T) {
-	var wg util.WaitGroupWrapper
-
 	testLock := func(rc bool, key string, tableName string) {
 		store, clean := testkit.CreateMockStore(t)
 		defer clean()
@@ -783,6 +782,7 @@ func TestPointGetLockExistKey(t *testing.T) {
 		))
 	}
 
+	var wg sync.WaitGroup
 	for i, one := range []struct {
 		rc  bool
 		key string
@@ -792,10 +792,12 @@ func TestPointGetLockExistKey(t *testing.T) {
 		{rc: true, key: "primary key"},
 		{rc: true, key: "unique key"},
 	} {
+		wg.Add(1)
 		tableName := fmt.Sprintf("t_%d", i)
-		wg.Run(func() {
-			testLock(one.rc, one.key, tableName)
-		})
+		go func(rc bool, key string, tableName string) {
+			defer wg.Done()
+			testLock(rc, key, tableName)
+		}(one.rc, one.key, tableName)
 	}
 	wg.Wait()
 }