<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #33750

Problem Summary:

### What is changed and how it works?

Note the closured value, it's a common trap for the newbies.

### Check List

Tests <!-- At least one of them must be included. -->

- [ ] Unit test
- [ ] Integration test
- [X] Manual test (add detailed scripts or steps below)

```
make ut X='run executor TestPointGetLockExistKey --race'
```

Before the fix:

```
--- FAIL: TestPointGetLockExistKey (3.58s)
    testing.go:1312: race detected during execution of test
FAIL
run all tasks takes 4.658048461s
make: *** [Makefile:121: ut] Error 1
```

After:

```
% make ut X='run executor  TestPointGetWriteLock --race'
tools/bin/ut run executor  TestPointGetWriteLock --race || { find $PWD/ -type d | grep -vE "(\.git|tools)" | xargs tools/bin/failpoint-ctl disable; exit 1; }
2022/04/11 15:38:01 maxprocs: Leaving GOMAXPROCS=16: CPU quota undefined
building task finish, count=1, takes=8.556764698s
run all tasks takes 992.22108ms
```

- [ ] No code

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```